+++ pkg/a2l/sharedextension.go
@@ -47,6 +47,18 @@ func (n *LongType) A2LString() string {
 	return fmt.Sprintf(format, n.Value)
 }
 
+func (n *ULongType) A2LString() string {
+	var format string
+
+	if n.Base == 10 {
+		format = fmt.Sprintf("%%0%vd", n.Size)
+	} else {
+		format = fmt.Sprintf("0x%%0%vX", n.Size)
+	}
+
+	return fmt.Sprintf(format, n.Value)
+}
+
 func (n *StringType) A2LString() string {
 	return fmt.Sprintf("\"%v\"", n.Value)
 }